Move OpenAI classifier to intermediate layer
