+++ src/adapters/adapter_openai.hpp
@@ -12,10 +12,9 @@ namespace adapters {
 class OpenAI: public CurlBase {
 public:
     OpenAI(const Configurations &configs);
-    std::expected<OpenAIClassification, OpenAIError> classify_instructions(const std::string &prompt);
     std::expected<OpenAIEdit, OpenAIError> query_edit_code(const std::string &prompt);
 
-private:
+protected:
     std::string query_responses_api_(const std::string &post_fields);
 
     std::string model_;
